ml_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
faiss-cpu==1.7.4
sentence-transformers==2.2.2
pydantic==2.5.0
numpy==1.24.3
python-multipart==0.0.6
